\name{kern.estJ}
\alias{kern.estJ}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Computes kernel density estimate
}
\description{
Computes kernel density estimate
}
\usage{
kern.estJ(sb.arg, band.h, sa.vec, ya.vec)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sb.arg}{
surrogate marker from Study B
}
  \item{band.h}{
bandwidth
}
  \item{sa.vec}{
surrogate marker from Study A
}
  \item{ya.vec}{
primary outcome from Study A
}
}

\value{
kernel density estimate
}

\author{
Jay Bartroff
}
\keyword{internal}
