\name{Kern.FUN}
\alias{Kern.FUN}
\title{
Calculates kernel matrix
}
\description{
Helper function; this calculates the kernel matrix
}
\usage{
Kern.FUN(zz, zi, bw)
}
\arguments{
  \item{zz}{
zz
}
  \item{zi}{
zi
}
  \item{bw}{
bandwidth
}
}
\value{
the kernel matrix
}
\author{
Layla Parast
}
\keyword{smooth}
\keyword{robust}
\keyword{nonparametric}
\keyword{internal}
