\name{VTM}
\alias{VTM}
\title{
Repeats a row.
}
\description{
Helper function; this function creates a matrix that repeats vc, dm times where each row is equal to the vc vector. 
}
\usage{
VTM(vc, dm)
}
\arguments{
  \item{vc}{
the vector to repeat.
}
  \item{dm}{
number of rows.
}
}
\value{
a matrix that repeats vc, dm times where each row is equal to the vc vector
}
\keyword{array}
\keyword{internal}